requirements.txt
imageio==2.34.2
joblib==1.4.2
lazy_loader==0.4
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
threadpoolctl==3.5.0
tifffile==2024.7.24
